requirements.txt
flask>=3.0.0
line-bot-sdk>=3.12.0
yfinance>=0.2.54
pandas>=2.2.0
numpy>=1.26.0
python-dotenv>=1.0.1
requests>=2.32.0
gunicorn>=22.0.0
